[IOS-9382] Remove ClipboardContent from Bonus Data
Removed clipboard content from the -allSystemInformation method (it’s no longer bundled into the overall dictionary of system info). The SystemServices code still allows you to specifically ask for clipboard content, but the way TN app uses this code, it won’t access clipboard content when it calls -allSystemInformation. Also added a little more code and a couple asserts to the demo app to confirm that we aren’t including clipboard content in the overal dictionary.